move prometheus_metrics to official public API
Moved prometheus metrics from internal _prometheus_metrics module to
official dynamo.prometheus_metrics public API. Updated imports in
example files and type stubs to use the new public API location.
